Rename from 'nsgenerator' to 'namespace'
